+++ web/src/modules/travel-authorizations/components/my-travel-authorizations-edit-estimate-page/EstimatesTable.vue
